+++ src/main/java/org/opensearch/searchrelevance/common/PluginConstants.java
@@ -30,6 +30,7 @@ public class PluginConstants {
     public static final String REMOTE_SEARCH_CONFIGURATIONS_URL = SEARCH_RELEVANCE_BASE_URI + "/remote_search_configurations";
     /** The URI for this plugin's remote search execute rest action */
     public static final String REMOTE_SEARCH_EXECUTE_URL = SEARCH_RELEVANCE_BASE_URI + "/remote_search/execute";
+
     /** The URI for initializing the UBI indices */
     public static final String INITIALIZE_URL = "/_plugins/ubi/initialize";