+++ CHANGELOG.md
@@ -11,8 +11,10 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Enhancements
 
 ### Bug Fixes
+* Updated ImportJudgmentsProcessor to handle ratings in numeric as well as string formats ([#230](https://github.com/opensearch-project/search-relevance/pull/230))
 
 ### Infrastructure
+* Updated System.env syntax for Gradle 9 compatibility ([#227](https://github.com/opensearch-project/search-relevance/pull/227))
 
 ### Documentation